ENH: hide toParallel
